Mantel-Haenszel pooled estimates of the prevalence ratio (PR) or the prevalence difference (PD) in the WASH Benefits trials

Mantel-Haenszel pooled estimates of the prevalence ratio (PR) or the prevalence difference (PD) using randomization block as the stratification variable.

washb_mh(Y,tr,strat,contrast,measure="RR")

Arguments

Y
binary outcome variable (here: diar7d)

tr
binary treatment group variable, comparison group first

strat
stratification variable (here: block)

contrast
vector of length 2 that includes the tr groups to contrast (control(reference arm) and then intervention)

measure
measure of effect. RR = prev ratio, RD = prev difference

Value

res: Matrix of RD, se.RD, ci.lb, ci.ub, Z, p-value.

Details

The function calls the M-H estimator for two different arms of the study. It relies on the rma.mh() function in the metafor package.

Estimate the Mantel-Haenszel prevalence ratio note: strata with no outcomes (i.e., missing PR) are dropped. This is consistent with a fixed-effects regression analysis, in which those strata would not contribute to the estimates. The arguments Y,tr,strat, below need to be from the same dataset.

Examples

#Prescreen function applied to the Bangladesh diarrheal disease outcome.
#The function will test a matrix of covariates and return those related to child diarrheal disease with
#a <0.2 p-value from a likelihood ratio test.

#Load  diarrhea data
library(washb)
data(washb_bd_diar)

#> Warning: data set 'washb_bd_diar' not found

data(washb_bd_enrol)

#> Warning: data set 'washb_bd_enrol' not found

 # drop svydate and month because they are superceded in the child level diarrhea data
washb_bd_enrol$svydate <- NULL

#> Error in washb_bd_enrol$svydate <- NULL: object 'washb_bd_enrol' not found

washb_bd_enrol$month <- NULL

#> Error in washb_bd_enrol$month <- NULL: object 'washb_bd_enrol' not found

# merge the baseline dataset to the follow-up dataset
ad <- merge(washb_bd_enrol,washb_bd_diar,by=c("dataid","clusterid","block","tr"),all.x=F,all.y=T)

#> Error in merge(washb_bd_enrol, washb_bd_diar, by = c("dataid", "clusterid",     "block", "tr"), all.x = F, all.y = T): object 'washb_bd_enrol' not found

# subset to the relevant measurement
# Year 1 or Year 2
ad <- subset(ad,svy==1|svy==2)

#> Error in subset(ad, svy == 1 | svy == 2): object 'ad' not found

#subset the diarrhea to children <36 mos at enrollment
### (exlude new births that are not target children)
ad <- subset(ad,sibnewbirth==0)

#> Error in subset(ad, sibnewbirth == 0): object 'ad' not found

ad <- subset(ad,gt36mos==0)

#> Error in subset(ad, gt36mos == 0): object 'ad' not found

# Exclude children with missing data
ad <- subset(ad,!is.na(ad$diar7d))

#> Error in subset(ad, !is.na(ad$diar7d)): object 'ad' not found

#Re-order the tr factor for convenience
ad$tr <- factor(ad$tr,levels=c("Control","Water","Sanitation","Handwashing","WSH","Nutrition","Nutrition + WSH"))

#> Error in factor(ad$tr, levels = c("Control", "Water", "Sanitation", "Handwashing",     "WSH", "Nutrition", "Nutrition + WSH")): object 'ad' not found

###Create vector of contrasts for each hypothesis to facilitate comparisons between arms.
#Hypothesis 1: Each intervention arm vs. Control
h1.contrasts <- list(
  c("Control","Water"),
  c("Control","Sanitation"),
  c("Control","Handwashing"),
  c("Control","WSH"),
  c("Control","Nutrition"),
  c("Control","Nutrition + WSH")
)

#Apply washb_mh to the water vs. control arm contrast.
washb_mh(Y=ad$diar7d,tr=ad$tr, contrast=c("Control","Water"), strat=ad$block,measure="RR")

#> Loading required package: metafor

#> Warning: package 'metafor' was built under R version 3.4.4

#> Loading required package: Matrix

#> Loading 'metafor' package (version 2.0-0). For an overview 
#> and introduction to the package please type: help(metafor).

#> Error in data.frame(Y = Y[tr == contrast[1] | tr == contrast[2]], tr = tr[tr ==     contrast[1] | tr == contrast[2]], strat = strat[tr == contrast[1] |     tr == contrast[2]]): object 'ad' not found

#Return the risk difference instead of the risk ration:
washb_mh(Y=ad$diar7d,tr=ad$tr, contrast=c("Control","Water"), strat=ad$block,measure="RD")

#> Error in data.frame(Y = Y[tr == contrast[1] | tr == contrast[2]], tr = tr[tr ==     contrast[1] | tr == contrast[2]], strat = strat[tr == contrast[1] |     tr == contrast[2]]): object 'ad' not found

#Use sapply command to efficiently apply the function to all the treatment arm contrasts

diff.h1 <- t(sapply(h1.contrasts,washb_mh,Y=ad$diar7d,tr=ad$tr,strat=ad$block,measure="RR"))

#> Error in data.frame(Y = Y[tr == contrast[1] | tr == contrast[2]], tr = tr[tr ==     contrast[1] | tr == contrast[2]], strat = strat[tr == contrast[1] |     tr == contrast[2]]): object 'ad' not found

rownames(diff.h1) <- c("Water v C","Sanitation v C","Handwashing v C","WSH v C","Nutrition v C","Nutrition + WSH v C")

#> Error in rownames(diff.h1) <- c("Water v C", "Sanitation v C", "Handwashing v C",     "WSH v C", "Nutrition v C", "Nutrition + WSH v C"): object 'diff.h1' not found

print(diff.h1)

#> Error in print(diff.h1): object 'diff.h1' not found